chore: tweak the return type of `dataEngine.query`
