+++ src/types/index.ts
@@ -1,15 +1,14 @@
 /* ONLY PLACE GENERAL TYPES HERE WHICH ARE USED THROUGHOUT THE APP
  * Types exported from here can be imported as follows:
  * `import type { MyType } from '@types'` */
-import type { ContextType } from '@dhis2/app-service-data'
 import type { TransformedAppCachedData } from '../components/app-wrapper/app-cached-data-query-provider'
-import type { ResponseErrorReport } from '@api/parse-engine-error'
 /* We have an ESLint rule in place to prevent imports from
  * `src/types/dhis2-openapi-schemas` anywhere else in the codebase.
  * The reason for this is so that we can apply manual overrides
  * for generated types here, as we have done for `SystemSettings`.
  * Anything that is needed from the generated types should be explicitly exported here;
  * this list should not contain the types we override. */
+export type { DataEngine, QueryResult, MutationResult } from './data-engine'
 export type {
     EventVisualizationType,
     FavoriteStatistics,
@@ -42,19 +41,6 @@ export type SingleQuery = {
         string,
         number | string | boolean | Array<number | string | boolean>
     >
-}
-export type DataEngine = ContextType['engine']
-export type QueryResult = Awaited<ReturnType<DataEngine['query']>>
-export type MutationResult = {
-    httpStatus: string
-    httpStatusCode: number
-    status: string
-    response: {
-        uid: string
-        klass: string
-        errorReports: Array<ResponseErrorReport>
-        responseType: string
-    }
 }
 export type { AppStore, AppDispatch, RootState } from '@store/store'
 export type { UseMetadataStoreReturnValue as MetadataStore } from '../components/app-wrapper/metadata-provider'
